Add functionality for receiving items

+++ worlds/sly1/Client/Addresses.cs
@@ -8,7 +8,7 @@ namespace Sly1AP
 {
     public class Addresses
     {
-        public ulong RaleighKeys = 0x2027CAB4;
+        public int RaleighKeys = 0x2027CAB4;
         public int MuggshotKeys = 0x2027CF00;
         public int MzRubyKeys = 0x2027D34C;
         public int PandaKingKeys = 0x2027D798;
